build: fix ARMv7 warnings under Werror
Use portable format specifiers for size_t and time_t values on 32-bit Ubuntu 26.04, where time_t is 64-bit.

Route TLS driver nsd downcasts through explicit helpers so ARM GCC does not reject the intentional object-model casts with -Wcast-align.

Wrap only generated protobuf-c output with a local -Wcast-align diagnostic suppression so handwritten impstats code still builds under strict warnings.

+++ plugins/omsendertrack/omsendertrack.c
@@ -33,6 +33,7 @@
  */
 #include "config.h"
 #include "rsyslog.h"
+#include <inttypes.h>
 #include <stdio.h>
 #include <stdarg.h>
 #include <stdlib.h>
@@ -186,8 +187,8 @@ static rsRetVal addSender(instanceData *const pData,
     sender_stats_t *stat;
     DEFiRet;
 
-    DBGPRINTF("addSender: Sender: %s, Messages: %" PRIu64 ", FirstSeen: %ld, LastSeen: %ld\n", sender, nMsgs, firstSeen,
-              lastSeen);
+    DBGPRINTF("addSender: Sender: %s, Messages: %" PRIu64 ", FirstSeen: %" PRIdMAX ", LastSeen: %" PRIdMAX "\n", sender,
+              nMsgs, (intmax_t)firstSeen, (intmax_t)lastSeen);
 
     CHKmalloc(stat = calloc(1, sizeof(sender_stats_t)));
     stat->sender = (const uchar *)strdup((const char *)sender);
